requirements.txt
openai>=1.14.3
transformers>=4.5.0
tiktoken>=0.5.0
nltk>=3.8.1
# Web
pyyaml>=5.4.1
plyvel==1.5.1
# Language detection for multilingual chunking
lingua-language-detector>=2.0.0
fastapi==0.115.3
aiohttp>=3.9.0
uvicorn[standard]==0.32.0
python-multipart==0.0.16
celery==5.4.0
redis==5.2.0
asgiref==3.8.1
tenacity==9.0.0

# Database and ORM
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Configuration and Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Security and Encryption
cryptography>=41.0.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
aiosqlite>=0.19.0  # For SQLite in-memory async tests

# Diff computation for version history
diff-match-patch>=20230430

# Document generation
python-docx>=0.8.11
# Markdown to DOCX conversion (LibreOffice for PDF)
markdown>=3.5
htmldocx>=0.0.6

# CLI tools for load testing
click>=8.0
rich>=13.0
websockets>=11.0
protobuf>=4.21.0
sentencepiece>=0.1.99
